requirements.txt
fasttext==0.9.3
langdetect==1.0.9
langid==1.1.6
lingua-language-detector==2.0.2
matplotlib==3.8.3
pandas==2.2.2
plotly==5.19.0
pymongo==4.8.0
PyMuPDF==1.24.9
seaborn==0.13.2
streamlit==1.38.0
torch==2.2.2
transformers==4.41.1
